pyproject.toml
[project]
name = "crewai_docs_updater"
version = "0.1.0"
description = "crewai-docs-updater using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "pygithub>=2.6.1",
]

[project.scripts]
kickoff = "crewai_docs_updater.main:kickoff"
run_crew = "crewai_docs_updater.main:kickoff"
plot = "crewai_docs_updater.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
